+++ sync_deploy/__init__.py
@@ -1,6 +1,8 @@
 #!usr/bin/env python
 # coding: utf-8
 
+from __future__ import unicode_literals
+
 __author__ = 'kylinfish@126.com'
 __date__ = '2013.8.15'
 __doc__ = u'自动化部署方案'