+++ sync_deploy/management/commands/compile.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+from __future__ import unicode_literals, print_function
+
 u"""编译项目源代码文件到pyc.
 """