+++ CHANGELOG.rst
@@ -1,7 +1,23 @@
 Changelog
 =========
 
-1.2 - 16-Feb-2015
+1.4 - 12-Apr-2015
+-----------------
+
+* Add logging info config
+* Fix fabric called may be error
+* Fix some text error
+* Refactor some code
+
+
+1.3 - 13-Mar-2015
+-----------------
+
+* Refactor some code
+* Add deploy record that use db
+
+
+1.2 - 07-Mar-2015
 -----------------
 
 * First release